2007-08-10 Farid Zaripov <Farid_Zaripov@epam.com>
	* _mutex.h (__rw_atomic_preincrement) [_WIN32]: Cast __x to int& instead of long&
	to use InterlockedIncrement() optimized function instead of generic one using
	critical sections.


git-svn-id: https://svn.apache.org/repos/asf/incubator/stdcxx/trunk@564698 13f79535-47bb-0310-9956-ffa450edef68

+++ include/rw/_mutex.h
@@ -1371,7 +1371,7 @@ __rw_atomic_preincrement (int &__x, bool)
 inline unsigned int
 __rw_atomic_preincrement (unsigned int &__x, bool)
 {
-    return __rw_atomic_preincrement (_RWSTD_REINTERPRET_CAST (long&, __x),
+    return __rw_atomic_preincrement (_RWSTD_REINTERPRET_CAST (int&, __x),
                                      false);
 }
